requirements.txt
altair==4.2.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==36.0.2
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.10.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.11.0
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
olefile==0.46
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydeck==0.7.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.1.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.8.1
systemd-python==234
tenacity==8.0.1
terminado==0.13.3
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
ttystatus==0.38
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
validators==0.18.2
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wsproto==1.1.0
xdg==5
xkit==0.0.0
zipp==1.0.0
